migrate manifests to use tgz utility
